requirements.txt
Django==4.2
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
